admin notification'

+++ app/Http/Controllers/userController.php
@@ -85,39 +85,48 @@ class userController extends Controller
             'role'=>'required'
 
         ]);
-     
-        $user=new User();
-        $user->firstname=$request->firstname;
-        $user->lastname=$request->lastname;
-        $user->email=$request->email;
-        $user->password=bcrypt($request->password);
-        $user->status=$request->status;
-        $user->role_id=$request->role;
-        $v=$user->save();
-        if($v)
+        if(User::where('email','=',$request->email)->where('role_id','=',$request->role)->count()>0)
         {
-            $user1= DB::table('users')->where('id', DB::raw("(select max(`id`) from users)"))->get();
-            foreach($user1 as $us)
-            {
-                $role=$us->role_id;
-                $id=$us->id;
-                $permissions= DB::table('role_has_permissions')->where(function ($query) use ($role,$id){
-                    $query->where('role_id','=', $role)->pluck('permission_id');
-                    // $query->where('id','=', $id);
-                    })->get();
-                $modelRole= DB::insert('insert into model_has_roles(role_id,model_type,model_id) values(?,?,?)',[$role,'App\User',$id]);
-                    
-                    foreach ($permissions as $per) {
-                    $permission= $per->permission_id;
+          Session::flash('alert-danger', 'Email id already exist!');
+          return redirect('admin/users'); 
+        }
+        else
+        {
+     
+          $user=new User();
+          $user->firstname=$request->firstname;
+          $user->lastname=$request->lastname;
+          $user->email=$request->email;
+          $user->password=bcrypt($request->password);
+          $user->status=$request->status;
+          $user->role_id=$request->role;
+          $user->created_date = date("Y-m-d H:i:s");
+          $v=$user->save();
+          if($v)
+          {
+              $user1= DB::table('users')->where('id', DB::raw("(select max(`id`) from users)"))->get();
+              foreach($user1 as $us)
+              {
+                  $role=$us->role_id;
+                  $id=$us->id;
+                  $permissions= DB::table('role_has_permissions')->where(function ($query) use ($role,$id){
+                      $query->where('role_id','=', $role)->pluck('permission_id');
+                      // $query->where('id','=', $id);
+                      })->get();
+                  $modelRole= DB::insert('insert into model_has_roles(role_id,model_type,model_id) values(?,?,?)',[$role,'App\User',$id]);
+                      
+                      foreach ($permissions as $per) {
+                      $permission= $per->permission_id;
 
-                    $modelPermissions= DB::insert('insert into model_has_permissions(permission_id,model_type,model_id) values(?,?,?)',[$permission,'App\User',$id]);                        
-                    }           
-            }
-            if($modelPermissions && $modelRole )
-            {
-                Session::flash('alert-success', 'User added!');
-                return redirect('admin/users');
-            }            
+                      $modelPermissions= DB::insert('insert into model_has_permissions(permission_id,model_type,model_id) values(?,?,?)',[$permission,'App\User',$id]);                        
+                      }           
+              }
+              if($modelPermissions && $modelRole )
+              {
+                  Session::flash('alert-success', 'User added!');
+                  return redirect('admin/users');
+              }            
+          }
         }
     }
 
@@ -164,40 +173,49 @@ class userController extends Controller
             'status'=>'required',
             'role'=>'required'
         ]);
-        $user = User::findOrFail($id);
-        $user->firstname=$request->firstname;
-        $user->lastname=$request->lastname;
-        $user->email=$request->email;
-        $user->status=$request->status;
-        $user->role_id=$request->role;
-        $user->created_date = date("Y-m-d H:i:s");
-        $user->save();
-        $modelPermissions;
-        if($user)
-        {
-            $role=$request->role;
-            $permissions= DB::table('role_has_permissions')->where(function ($query) use ($role,$id){
-                    $query->where('role_id','=', $role)->pluck('permission_id');
-                    // $query->where('id','=', $id);
-                    })->get();
-            $modelRole= DB::update('update model_has_roles set role_id=? where model_id=?',[$role,$id]);
-            $deletePermission=DB::delete('delete from model_has_permissions where model_id=?',[$id]);
-            if($deletePermission)
-                    {
-                        foreach ($permissions as $per) {
-                        $permission= $per->permission_id;
-                        $modelPermissions= DB::insert('insert into model_has_permissions(permission_id,model_type,model_id) values(?,?,?)',[$permission,'App\User',$id]);
-                        }       
-                    }                    
-       
-            if($modelPermissions && $modelRole)
-            {
-                Session::flash('alert-success', 'User updated!');
 
-                return redirect('admin/users');
+        if(User::where('email','=',$request->email)->where('role_id','=',$request->role)->count()>0)
+        {
+          Session::flash('alert-danger', 'Email id already exist!');
+          return redirect('admin/users'); 
+        }
+        else
+        {
+     
+            $user = User::findOrFail($id);
+            $user->firstname=$request->firstname;
+            $user->lastname=$request->lastname;
+            $user->email=$request->email;
+            $user->status=$request->status;
+            $user->role_id=$request->role;
+            $user->created_date = date("Y-m-d H:i:s");
+            $user->save();
+            $modelPermissions;
+            if($user)
+            {
+                $role=$request->role;
+                $permissions= DB::table('role_has_permissions')->where(function ($query) use ($role,$id){
+                        $query->where('role_id','=', $role)->pluck('permission_id');
+                        // $query->where('id','=', $id);
+                        })->get();
+                $modelRole= DB::update('update model_has_roles set role_id=? where model_id=?',[$role,$id]);
+                $deletePermission=DB::delete('delete from model_has_permissions where model_id=?',[$id]);
+                if($deletePermission)
+                        {
+                            foreach ($permissions as $per) {
+                            $permission= $per->permission_id;
+                            $modelPermissions= DB::insert('insert into model_has_permissions(permission_id,model_type,model_id) values(?,?,?)',[$permission,'App\User',$id]);
+                            }       
+                        }                    
+           
+                if($modelPermissions && $modelRole)
+                {
+                    Session::flash('alert-success', 'User updated!');
+
+                    return redirect('admin/users');
+                }
             }
         }
-            
     }
     /**
      * Remove the specified resource from storage.
@@ -215,20 +233,30 @@ class userController extends Controller
 
     public function updateMyInfo(Request $request)
     {
-       $validate=$request->validate([
-       "firstname"=>"required",
-       "lastname"=>"required",
-       "email"=>"required|email",
-       ]);
-       $firstname='';
-       $lastname='';
-       $email='';
-       foreach(User::where("id",'=',Auth::user()->id)->get() as $user)
-       {
-        $firstname=$user->firstname;
-        $lastname=$user->lastname;
-        $email=$user->email;
-       }
+        $validate=$request->validate([
+        "firstname"=>"required",
+        "lastname"=>"required",
+        "email"=>"required|email",
+        ]);
+        $firstname='';
+        $lastname='';
+        $email='';
+        foreach(User::where("id",'=',Auth::user()->id)->get() as $user)
+        {
+            $firstname=$user->firstname;
+            $lastname=$user->lastname;
+            $email=$user->email;
+        }
+        if($email != $request->email)
+        {
+            $isEmailExist=User::where('email','=',$request->email)->get();
+            if(count($isEmailExist)>0)
+            {
+                Session::flash('alert-danger', 'This email id is already registered!! Use another email Id');
+                return redirect()->back();
+
+            }
+        }
 
        $oldEmail=Auth::user()->email;       
        if($firstname!=$request->firstname)
